replace some float with double
git-svn-id: http://svn.cern.ch/guest/lhcb/Hlt/trunk@81483 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: d6b3e609e7d3679024afc70316be9492088b7893

+++ Hlt/Hlt/HltMuon/src/HltMuonRec.h
@@ -1,4 +1,4 @@
-// $Id: HltMuonRec.h,v 1.7 2009-04-29 14:18:11 asatta Exp $
+// $Id: HltMuonRec.h,v 1.8 2010-02-12 07:49:12 graven Exp $
 #ifndef HLTMUON_H 
 #define HLTMUON_H 1
 
@@ -47,7 +47,7 @@ public:
   StatusCode detectClone();    
   StatusCode strongCloneKiller();
 
-  void insertMatch(int muon,LHCb::Track* track,float distance);
+  void insertMatch(int muon,LHCb::Track* track,double distance);
 	StatusCode printOut();  
   StatusCode getPads(int station);
   StatusCode createCoordsFromLC(int station);
